Award criteria, scoring and thresholds
Award criteria, scoring and thresholds are described in Annex D of the Work Programme General Annexes(wp-13-general-annexes_horizon-2021-2022_en.pdf), Page 18:  Research and innovation actions (RIA)
RIA is for "Establish new knowledge and/or explore the feasibility of a new or improved technology"

Scores and weighting (page 21)
Horizon Europe - Work Programme 2021-2022 General Annexes
 Evaluation scores will be awarded for the criteria, and not for the different aspects listed in the table. For full applications, each criterion will be scored out of 5. The threshold for individual criteria will be 3. The overall threshold, applying to the sum of the three individual scores, will be 10.
To determine the ranking for ‘Innovation actions’, the score for ‘Impact’ will be given a weight of 1.5.
Proposals that pass the individual threshold AND the overall threshold will be considered for funding, within the limits of the available call budget. Other proposals will be rejected.