Dissemination ande exploitation plan deliverable.
from the commentary ```In case your proposal is selected for
funding, a more detailed plan for dissemination and exploitation
including communication activities will need to be provided as a
mandatory project deliverable within 6 months after signature date.
This plan shall be periodically updated in alignment with the projects
progress. ```

Chosen 3 month delivery due date as we may sign before the project starts.