first review with suggested changes
To note: I introduced 'paragraphs' where they were removed. We need those to get
section numbers, otherwise we cannot refer to the labels of those sections. (We
can, but the numbers are wrong: they refer to the last section defined before
than label is in the text.)

I am out of time now, but I think the attached is already nearly good enough to merge.